structure: split modules into files
